refactor: 즐겨찾기 탭의 상태와 로직을 커스텀 훅으로 분리
